+++ src/App/VRChatContentPublisher.App/Extensions/HostBuilderExtension.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Trace;
@@ -6,6 +7,7 @@ using VRChatContentPublisher.BundleProcessCore.Telemetry;
 using VRChatContentPublisher.ConnectCore.Telemetry;
 using VRChatContentPublisher.Core.Telemetry;
 using VRChatContentPublisher.PersistentCore.Telemetry;
+using VRChatContentPublisher.TelemetryCore;
 using VRChatContentPublisher.TelemetryCore.Extensions;
 using VRChatContentPublisher.VRChatApi.Telemetry;
 
@@ -18,6 +20,14 @@ public static class HostBuilderExtension
         string appSessionLifetimeId
     ) where T : IHostApplicationBuilder
     {
+        builder.Configuration.AddInMemoryCollection(new Dictionary<string, string?>
+        {
+            // https://github.com/open-telemetry/opentelemetry-dotnet/blob/d141db4406ed556e8c57cdef6d7a09bf51295bf4/src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/ExperimentalOptions.cs#L36-L48
+            // https://github.com/open-telemetry/opentelemetry-dotnet/blob/1fafead47395a517ff827d61cef731052849f90f/src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpExporterOptionsExtensions.cs#L104-L116
+            { "OTEL_DOTNET_EXPERIMENTAL_OTLP_RETRY", "disk" },
+            { "OTEL_DOTNET_EXPERIMENTAL_OTLP_DISK_RETRY_DIRECTORY_PATH", TelemetryConst.GetOtlpDiskRetryPath() }
+        });
+
         builder.Services.AddOpenTelemetry()
             .WithTracing(tracing =>
             {